numpy data type is fixed.  - should be int16.!!

+++ APP/_InterpretationViewer/blocks/DBMManager.py
@@ -200,7 +200,7 @@ class DBMManager(QObject):
         vtk_img.SetSpacing(s)
         # # vtk_img.SetExtent(e)
         vtk_img.AllocateScalars(vtk.VTK_CHAR, 1)
-        buf = np.zeros(np.product(dim))
+        buf = np.zeros(np.product(dim)).astype(np.int16)     # NOTE make sure np.int16 type!
         buf.fill(-1000)
         vtk_img.GetPointData().SetScalars(dsa.numpyTovtkDataArray(buf))
         vtk_img_narray = buf.reshape(dim, order='F')
